Added user authentication

+++ comments/views.py
@@ -1,15 +1,24 @@
 # Create your views here.
 from django.contrib import messages
 from django.contrib.contenttypes.models import ContentType
-from django.http import HttpResponseRedirect
-from django.shortcuts import render, get_object_or_404
+from django.http import HttpResponseRedirect, Http404, HttpResponse
+from django.shortcuts import render
 
 from comments.forms import CommentForm
 from comments.models import Comment
 
 
 def comment_delete(request, id):
-    obj = get_object_or_404(Comment, id=id)
+    try:
+        obj = Comment.objects.get(id=id)
+    except:
+        raise Http404
+
+    if obj.user != request.user:
+        response = HttpResponse('You do not have permission to do this')
+        response.status_code = 403
+        return response
+
     if request.method == 'POST':
         parent_object_url = obj.content_object.get_absolute_url()
         obj.delete()
@@ -22,7 +31,16 @@ def comment_delete(request, id):
 
 
 def comment_thread(request, id):
-    obj = get_object_or_404(Comment, id=id)
+    try:
+        obj = Comment.objects.get(id=id)
+    except:
+        raise Http404
+
+    if not obj.is_parent:
+        obj = obj.parent
+
+
+
     initial_data = {
         'content_type': obj.content_type,
         'object_id': obj.object_id
